Return serverId for backups in Rackspace driver
Backups on the Rackspace Cloud are stored as images, so they can be seen
in response to list_images. However, as they are generically named, it's
currently impossible to tell of which Node they're a backup.

The Rackspace Cloud API returns the serverId of the origin Node. Add
this to the extra dict of the image.